+++ lib/mx_app_icon.dart
@@ -1,2 +1,3 @@
 export 'generator.dart';
 export 'util.dart';
+export 'color_parse.dart';